requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-MySQL==1.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
PyMySQL==1.1.1
Werkzeug==3.1.3
